# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Bible Atlas API
  description: The Bible Atlas HTTP API, generated from the Rust that serves it.
  version: 0.8.0
paths:
...
  /api/chapter/{cref}:
    get:
      tags:
      - reading
      summary: |-
        `GET /api/chapter/{cref}`. `cref` must parse as exactly a
        `ScriptureRef::Chapter` (book + chapter, e.g. `EXO.14`) — a book-only or
        verse/passage-shaped path segment is the wrong shape for this endpoint
        and 400s as `bad_ref`, same as an unparseable one. The optional
        `?translation=kjv` query param (ruling 5, M1 is KJV-only) is never
        extracted, so its presence or absence cannot affect this handler at all.
      description: |-
        ruling-3-policy: once `cref` parses as a `Chapter`, an out-of-range
        chapter number (or a book with no known chapters in this atlas) is NOT an
        error — the verse-count bound comes from `canon.books[].chapters`, and an
        unknown/short chapter just yields `verse_count = 0`, i.e. a 200 response
        with an empty `verses` list. Same rationale as `scene_scripture`: a
        reader showing "no verses in this chapter" is a meaningful response, not
        a failure.

        Batch M-A (brief requirement 5, "re-implement the OLD /api/chapter
        handler as a VIEW over the window query"): the verse TEXT below now
        comes from `GraphState::chapter_span` + `GraphState::window` -- the SAME
        windowed reading-order query `GET /api/text?scope=chapter` calls --
        instead of `data.verses.get(key)`. The verse-count bound and the
        out-of-canon policy above are UNCHANGED and still sourced from
        `AtlasData`; headings moved to `graph.heading_index` in M-C2, and
        OVERLAY-1 Task 5 moved the PLACE-MENTION half onto the port too
        (`graph.scene_source(&data)`'s own `places_for_verse`/`place`, the
        materialised-from-the-graph successors of the deleted
        `AtlasData::places_for_verse`/`place_by_id` -- identical ids in
        identical order, see those methods' own doc comments). THIS endpoint
        (the reader's own chapter view) was untouched by Batch M-B's own
        event-world migration;
        only `/api/narrative/event/{id}` (see that handler's own doc comment)
        and the generic `/api/node`/`/edges` endpoints move to the graph this
        batch. The WIRE SHAPE is byte-for-byte identical -- proven by
        `tests/graph_equivalence.rs`'s own all-1,189-chapters comparison -- so
        every existing caller of this endpoint (the reader's chapter view/
        mini-reader/split view, `ChapterNode`, `PlaceCard`'s hover verse text,
        `PassageBlock`, `PopoverSectionProviders`) now serves from the graph
        with NO client-side change and no reader-visible behavior change.
      operationId: chapter
      parameters:
      - name: cref
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
        '400':
          description: bad_window | bad_ref | bad_kind | bad_dir | bad_corpus
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
        '404':
          description: not_found
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
        '500':
          description: internal
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
...
  /api/polities:
    get:
      tags:
      - map
      summary: |-
        `GET /api/polities?from=&to=`. `from`/`to` share `scene_time`'s lenient
        parsing (ruling 1: missing/unparseable/zero/inverted -> 400
        `bad_window`), via the same `parse_year` helper and `TimeRange::new`
        validity check.
      description: |-
        Once the window is valid, emits every era (of every polity) whose own
        `[from,to]` intersects it -- a polity with several eras in view (a
        window spanning a border change) contributes one row PER intersecting
        era, all sharing that polity's own `id`/`color_key`; a window matching no
        era at all (out-of-span, or the `demo_fixture()`/pre-ETL case where
        `data.polities` is empty) is not an error, mirroring `scene_time`'s own
        ruling-2 spirit -- it 200s with an empty `polities` array. Deterministic
        order: by polity id, then by era `from` -- so a multi-era window always
        lists a polity's OLDER era before its newer one (the exact order
        map.js's `BorderLayer` needs to paint older-under-newer and pick the
        dotted/lightest era correctly without re-sorting client-side).
        BATCH M-C (controller decision 7): re-implemented as a VIEW over the
        graph's own Polity nodes (`atlas_graph::polity_adapter`) instead of
        `AtlasData.polities` -- border data as node payloads (controller
        decision 2), the map consuming payloads directly, per era, off each
        Polity node's own `NodePayload::Polity.eras`. `GraphService.polity_ids`
        is the companion enumeration (same status as `era_ids`, see that
        field's own doc comment). WIRE SHAPE UNCHANGED: `Polity`/
        `PolityDelta`, same fields, same conditional presence, same
        deterministic sort -- `AtlasData.polities`/`polities.json` stay
        standing (deliberately NOT this batch's deletion target: still the
        adapter's own curated source, same status as `event_world`'s own
        `atlas.events`/`.narratives`).
      operationId: polities
      parameters:
      - name: from
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: to
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Polities'
        '400':
          description: bad_window | bad_ref | bad_kind | bad_dir | bad_corpus
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
        '404':
          description: not_found
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
        '500':
          description: internal
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
  /api/scene:
    get:
      tags:
      - map
      summary: |-
        `GET /api/scene?from=&to=`. `from`/`to` are read out of a
        `Query<HashMap<String, String>>` rather than a strongly-typed `Query<T>`
        specifically so a missing or unparseable value can never trigger axum's
        own extractor-rejection response (ruling 1) — `HashMap<String, String>`
        cannot itself fail to deserialize on these inputs, so every failure mode
        (missing, non-integer, zero, inverted) is handled by this function and
        always yields the typed `bad_window` body.
      description: |-
        OVERLAY-1 Task 5: the scene's DATA now comes from the graph port --
        `graph.scene_source(&data)`, an `atlas_graph::scene_source::
        GraphSceneSource` built once at load from `GraphService`'s own snapshot,
        not from `AtlasData`'s former graph-derived `events`/`places`/
        `narratives` fields (which `legacy::atlas_data_overlay` used to
        reconstruct at boot, and which are no longer written on any serving
        path). `data` is still extracted because that source reads two genuinely
        curated-JSON sidecars through it (`place-history.json`,
        `place-names-kjv.json`). The composed bytes are unchanged -- `tests/
        scene_byte_identity.rs`'s 25 pinned hashes are the gate on that.
      operationId: scene_time
      parameters:
      - name: from
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: to
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scene'
        '400':
          description: bad_window | bad_ref | bad_kind | bad_dir | bad_corpus
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
        '404':
          description: not_found
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
        '500':
          description: internal
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
  /api/scene/scripture:
    get:
      tags:
      - map
      summary: '`GET /api/scene/scripture?ref=`.'
      description: |-
        ruling-3-policy: a `ref` that fails to *parse* (unknown book code, zero
        chapter/verse, empty segment, inverted range — i.e. `ScriptureRef::parse`
        returns `Err`) is a structurally bad ref and always 400s as `bad_ref`,
        same as a missing `ref` param (treated as parsing the empty string, which
        also fails to parse). A `ref` that *parses* but names coordinates outside
        the loaded canon (e.g. a chapter number past the end of the book) is
        deliberately NOT an error: `compose_scripture_scene` only ever matches it
        against verses that actually exist on events/places, so an out-of-canon
        ref naturally composes an empty-but-valid scene, exactly mirroring
        ruling 2's "don't reject out-of-span time windows" for the sibling
        endpoint. This needs no extra bounds-checking code — it falls out of not
        adding any.

        OVERLAY-1 Task 5: composes from `graph.scene_source(&data)`, exactly as
        `scene_time` above does -- see that handler's own doc comment.
      operationId: scene_scripture
      parameters:
      - name: ref
        in: query
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scene'
        '400':
          description: bad_window | bad_ref | bad_kind | bad_dir | bad_corpus
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
        '404':
          description: not_found
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
        '500':
          description: internal
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
...
components:
  schemas:
...
    Chapter:
      type: object
      required:
      - ref
      - book
      - chapter
      - verses
      properties:
        book:
          type: string
        chapter:
          type: integer
          format: int32
          minimum: 0
        ref:
          type: string
        verses:
          type: array
          items:
            $ref: '#/components/schemas/Verse'
      additionalProperties: false
...
    Polities:
      type: object
      required:
      - polities
      properties:
        polities:
          type: array
          items:
            $ref: '#/components/schemas/Polity'
      additionalProperties: false
    Polity:
      type: object
      description: |-
        One polity-era row: `id` is the POLITY's (constant across every era row
        the same polity contributes to a response) and `color_key` its own
        precomputed hash, while `name`/`from`/`to`/`rings` are this specific ERA's.
        The field names deliberately match `atlas_core::data::PolityEra`'s own
        (`from`/`to`, not `from_year`/`to_year` -- see that struct's doc comment
        for why), so the era half is a plain copy, not a rename.
      required:
      - id
      - name
      - from
      - to
      - rings
      - color_key
      properties:
        color_key:
          type: integer
          format: int32
          minimum: 0
        fall:
          oneOf:
          - $ref: '#/components/schemas/PolityDelta'
          - type: 'null'
        from:
          type: integer
          format: int32
        id:
          type: string
        name:
          type: string
        rings:
          type: array
          items:
            type: array
            items:
              type: array
              items: false
              prefixItems:
              - type: number
                format: double
              - type: number
                format: double
        to:
          type: integer
          format: int32
        transition:
          oneOf:
          - $ref: '#/components/schemas/PolityDelta'
            description: |-
              Batch M requirement 1: this era's own delta metadata, OMITTED (not
              null) when absent -- "an uneventful boundary stays visible but gets
              the minimal popover," so the client must be able to tell "no
              transition curated" apart from "transition curated with zero
              verses" (the latter still shows the event/grounding-note sections,
              just no THE SCRIPTURES section — see `PolityDelta::verses`' own doc
              comment). Mirrors `NarrativePosition.prior`/`.following`'s own
              `skip_serializing_if` precedent exactly.
          - type: 'null'
      additionalProperties: false
...
    Scene:
      type: object
      required:
      - mode
      - places
      - quiet_places
      - arrows
      - narratives
      properties:
        arrows:
          type: array
          items:
            $ref: '#/components/schemas/SceneArrow'
        mode:
          $ref: '#/components/schemas/SceneMode'
        narratives:
          type: array
          items:
            $ref: '#/components/schemas/SceneNarrative'
        places:
          type: array
          items:
            $ref: '#/components/schemas/ScenePlace'
        quiet_places:
          type: array
          items:
            $ref: '#/components/schemas/QuietPlace'
          description: |-
            Batch E2 (the ever-present graph, user direction 2026-08-19: "all of
            the cities in our graph are available in any timerange rather than
            just loading those which are biblically active at the time"): every
            event-bearing place NOT in `places` for this window -- period
            relevance is now expressed by GLOW (whether a place is in `places`),
            not existence. QUIET-1: for every time-mode window, `places` union
            `quiet_places` equals the FULL event-bearing place set (fixed
            cardinality, derived from the data -- see
            `AtlasData::event_bearing_place_ids`), and the two are disjoint.
            ALWAYS an array (never omitted) on both modes -- but always EMPTY for
            scripture-mode scenes, which never gain quiet places at all (chosen
            over an `Option`/omitted-key representation for wire regularity with
            `places` itself, which is likewise never optional; a client reading
            `quiet_places` needs no null-check either way).
        ref:
          type:
          - string
          - 'null'
        window:
          oneOf:
          - $ref: '#/components/schemas/TimeRange'
          - type: 'null'
      additionalProperties: false
    SceneArrow:
      type: object
      required:
      - narrative
      - color
      - from_place
      - to_place
      - from_event
      - to_event
      - order
      properties:
        color:
          type: string
        from_event:
          type: string
        from_place:
          type: string
        narrative:
          type: string
        order:
          type: integer
          format: int32
          minimum: 0
        to_event:
          type: string
        to_place:
          type: string
      additionalProperties: false
...
    SceneMode:
      type: string
      description: |-
        Which question a scene answers: what was happening in a span of years, or
        where a passage happens. Every scene is composed by exactly one of
        `scene::compose_time_scene`/`compose_scripture_scene`, so the pair is
        closed by construction.
      enum:
      - time
      - scripture
    SceneNarrative:
      type: object
      required:
      - id
      - name
      - color
      - legs_in_scene
      properties:
        color:
          type: string
        id:
          type: string
        legs_in_scene:
          type: integer
          format: int32
          minimum: 0
        name:
          type: string
      additionalProperties: false
...
    Verse:
      type: object
      required:
      - verse
      - text
      - places
      - persons
      - xref_count
      - words_of_christ
      properties:
        heading:
          oneOf:
          - $ref: '#/components/schemas/Heading'
            description: |-
              Batch T requirement 5: this verse's own pericope heading, when it is
              a heading ANCHOR (`AtlasData::heading_for_verse`) -- omitted (not
              null), matching `Polity.transition`/`.fall`'s own conditional-
              presence convention, since the overwhelming majority of verses never
              anchor a heading at all.
          - type: 'null'
        persons:
          type: array
          items:
            $ref: '#/components/schemas/PersonRef'
          description: |-
            M-D3 (owner ruling U5): every person the graph's own `mentions`
            relation attests at this verse's own locus, in wire (row-insertion)
            order -- see `PersonRef`'s own doc comment for the full
            mentions-attested-only reasoning. Always present, possibly empty
            (most verses mention zero persons by name) -- SAME conditional-
            presence-lives-client-side convention `places` immediately above
            already establishes for this exact shape of field.
        places:
          type: array
          items:
            $ref: '#/components/schemas/PlaceRef'
          description: |-
            Batch R requirement 5: every curated place whose OWN `verse_links`
            names this verse, in place-id order (stable, not meaningful --
            there is no "primary mention" concept). Always present, possibly
            empty (most verses geocode to zero places) -- conditional presence
            lives on the CLIENT side (a mini-reader renders no hoverable span at
            all when this is empty), not as an omitted wire key, matching every
            other "always an array" field in this app's own wire (e.g.
            `Scene.quiet_places`).
        text:
          type: string
        verse:
          type: integer
          format: int32
          minimum: 0
        words_of_christ:
          type: array
          items:
            $ref: '#/components/schemas/WordsOfChristSpan'
          description: |-
            Batch RED-1: this verse's own aligned sub-verse red-letter spans
            (`GraphService.red_letter_spans`, precomputed once -- see that
            field's own doc comment for why this ONE companion is not
            graph-derivable), in ascending order. Always present, possibly
            empty -- SAME "always an array" convention `places`/`persons`
            above already establish (most verses carry none). KJV-ONLY by
            construction (decision 5: sub-verse precision is KJV-specific;
            this endpoint has never served any other translation -- ruling 5,
            "M1 is KJV-only").
        xref_count:
          type: integer
          description: |-
            Batch M-D2 (owner's cross-reference superscript directive, closed on
            the graph platform): this verse's own `cites` edge-summary COUNT at
            its TextUnit locus -- read straight off THE PORT (`GraphQuery::
            edge_summary`, design doc §5), the SAME generic query the `/api/node/
            {id}` handler itself calls (`graph::node_card`) -- never a
            second, bespoke tally. This is a genuinely NEW field (no bespoke
            predecessor existed), sourced NATIVELY from the generic contract from
            day one: the reader's own superscript markers ARE a query of the
            graph, per the design's own P1/P4 laws, just batched onto the
            EXISTING per-chapter response (one HTTP round trip for a whole
            chapter) rather than issuing 20-30 individual `GET /api/node/{id}`
            calls per chapter, which `client/IExplorableClient.cs`'s own concrete
            implementation is what a SINGLE node's lookup uses (see that file's
            own doc comment + `GraphExplorableClientTests` for the direct,
            unbatched proof the same contract is correct in isolation). Always
            present (0 for the overwhelming majority of verses, never omitted --
            the client's own superscript decision is `count == 0` -> nothing,
            never a missing-field special case).
          minimum: 0
      additionalProperties: false
...